requirements.txt
mysql_connector_repackaged==0.3.1
pandas==2.2.3
pymongo==4.10.1
python-dotenv==1.0.1
python-telegram-bot==21.7
